Adds ability to get rule alias from Rule object or class
This works with a simple flip + lookup of the `RulesMaped::$rules` array.

This means that the values of the `RulesMaped::$rules` array must be unique, otherwise there will be collisions when the array is flipped.

The `validates custom rule (required) rule successfully` test has been changed to implement a custom rule, rather than just