Ignores all other rules if `NullableRule` is present and the tested field is null.
Given an array of rules that include `nullable`, other rules should only be tested if the field exists in the data.

This commit also resolves all rule aliases to their respective `Rule` objects before passing them to the `checkPasses()` loop; this allows checking for the presence of `NullableRule` in the rule array.